+++ src/core-interfaces.ts
@@ -9,13 +9,13 @@ import { CHART_NAME, PLACEHOLDER_TYPE, SHAPE_NAME, SLIDE_OBJECT_TYPES, TEXT_HALI
 
 /**
  * Coordinate number - either:
- * - Inches
- * - Percentage
+ * - Inches (0-n)
+ * - Percentage (0-100)
  *
  * @example 10.25 // coordinate in inches
  * @example '75%' // coordinate as percentage of slide size
  */
-export type Coord = number | string
+export type Coord = number | `${number}%`
 export type PositionProps = {
 	/**
 	 * Horizontal position
@@ -269,12 +269,6 @@ export interface TextBaseProps {
 	 */
 	breakLine?: boolean
 	/**
-	 * Add a soft line-break (shift+enter) before line text content
-	 * @default false
-	 * @since v3.5.0
-	 */
-	softBreakBefore?: boolean
-	/**
 	 * Add standard or custom bullet
 	 * - use `true` for standard bullet
 	 * - pass object options for custom bullet
@@ -390,6 +384,12 @@ export interface TextBaseProps {
 	 */
 	lang?: string
 	/**
+	 * Add a soft line-break (shift+enter) before line text content
+	 * @default false
+	 * @since v3.5.0
+	 */
+	softBreakBefore?: boolean
+	/**
 	 * tab stops
 	 * - PowerPoint: Paragraph > Tabs > Tab stop position
 	 * @example [{ position:1 }, { position:3 }] // Set first tab stop to 1 inch, set second tab stop to 3 inches
@@ -477,24 +477,34 @@ export interface ImageProps extends PositionProps, DataOrPathProps, CommonProps 
 		type: 'contain' | 'cover' | 'crop'
 		/**
 		 * Image width
+		 * - inches or percentage
+		 * @example 10.25 // position in inches
+		 * @example '75%' // position as percentage of slide size
 		 */
-		w: number
+		w: Coord
 		/**
 		 * Image height
+		 * - inches or percentage
+		 * @example 10.25 // position in inches
+		 * @example '75%' // position as percentage of slide size
 		 */
-		h: number
+		h: Coord
 		/**
-		 * Area horizontal position related to the image
-		 * - Values: 0-n
+		 * Offset from left to crop image
 		 * - `crop` only
+		 * - inches or percentage
+		 * @example 10.25 // position in inches
+		 * @example '75%' // position as percentage of slide size
 		 */
-		x?: number
+		x?: Coord
 		/**
-		 * Area vertical position related to the image
-		 * - Values: 0-n
+		 * Offset from top to crop image
 		 * - `crop` only
+		 * - inches or percentage
+		 * @example 10.25 // position in inches
+		 * @example '75%' // position as percentage of slide size
 		 */
-		y?: number
+		y?: Coord
 	}
 }
 /**
@@ -507,6 +517,19 @@ export interface MediaProps extends PositionProps, DataOrPathProps, CommonProps 
 	 * - Use 'online' to embed a YouTube video (only supported in recent versions of PowerPoint)
 	 */
 	type: MediaType
+	/**
+	 * Cover image
+	 * @since 3.9.0
+	 * @default "play button" image, gray background
+	 */
+	cover?: string
+	/**
+	 * media file extension
+	 * - use when the media file path does not already have an extension, ex: "/folder/SomeSong"
+	 * @since 3.9.0
+	 * @default extension from file provided
+	 */
+	extn?: string
 	/**
 	 * video embed link
 	 * - works with YouTube
@@ -939,7 +962,6 @@ export interface TextPropsOptions extends PositionProps, DataOrPathProps, TextBa
 	glow?: TextGlowProps
 	hyperlink?: HyperlinkProps
 	indentLevel?: number
-	inset?: number
 	isTextBox?: boolean
 	line?: ShapeLineProps
 	/**
@@ -956,6 +978,15 @@ export interface TextPropsOptions extends PositionProps, DataOrPathProps, TextBa
 	 * @since v3.5.0
 	 */
 	lineSpacingMultiple?: number
+	// TODO: [20220219] powerpoint uses inches but library has always been pt... @future @deprecated - update in v4.0? [range: 0.0-22.0]
+	/**
+	 * Margin (points)
+	 * - PowerPoint: Format Shape > Shape Options > Size & Properties > Text Box > Left/Right/Top/Bottom margin
+	 * @default "Normal" margin in PowerPoint [3.5, 7.0, 3.5, 7.0] // (this library sets no value, but PowerPoint defaults to "Normal" [0.05", 0.1", 0.05", 0.1"])
+	 * @example 0 // Top/Right/Bottom/Left margin 0 [0.0" in powerpoint]
+	 * @example 10 // Top/Right/Bottom/Left margin 10 [0.14" in powerpoint]
+	 * @example [10,5,10,5] // Top margin 10, Right margin 5, Bottom margin 10, Left margin 5
+	 */
 	margin?: Margin
 	outline?: { color: Color; size: number }
 	paraSpaceAfter?: number
@@ -984,6 +1015,10 @@ export interface TextPropsOptions extends PositionProps, DataOrPathProps, TextBa
 	strike?: boolean | 'dblStrike' | 'sngStrike'
 	subscript?: boolean
 	superscript?: boolean
+	/**
+	 * Vertical alignment
+	 * @default middle
+	 */
 	valign?: VAlign
 	vert?: 'eaVert' | 'horz' | 'mongolianVert' | 'vert' | 'vert270' | 'wordArtVert' | 'wordArtVertRtl'
 	/**
@@ -994,7 +1029,7 @@ export interface TextPropsOptions extends PositionProps, DataOrPathProps, TextBa
 	wrap?: boolean
 
 	/**
-	 * Whather "Fit to Shape?" is enabled
+	 * Whether "Fit to Shape?" is enabled
 	 * @deprecated v3.3.0 - use `fit`
 	 */
 	autoFit?: boolean
@@ -1003,6 +1038,11 @@ export interface TextPropsOptions extends PositionProps, DataOrPathProps, TextBa
 	 * @deprecated v3.3.0 - use `fit`
 	 */
 	shrinkText?: boolean
+	/**
+	 * Inset
+	 * @deprecated v3.10.0 - use `margin`
+	 */
+	inset?: number
 	/**
 	 * Dash type
 	 * @deprecated v3.3.0 - use `line.dashType`
@@ -1077,14 +1117,14 @@ export interface IChartPropsBase {
 	border?: BorderProps
 	chartColors?: HexColor[]
 	/**
-	 * opacity (0.0 - 1.0)
-	 * @example 0.5 // 50% opaque
+	 * opacity (0 - 100)
+	 * @example 50 // 50% opaque
 	 */
 	chartColorsOpacity?: number
 	dataBorder?: BorderProps
 	displayBlanksAs?: string
 	fill?: HexColor
-	invertedColors?: string
+	invertedColors?: HexColor[]
 	lang?: string
 	layout?: PositionProps
 	shadow?: ShadowProps
@@ -1231,8 +1271,22 @@ export interface IChartPropsChartBar {
 	bar3DShape?: string
 	barDir?: string
 	barGapDepthPct?: number
+	/**
+	 * MS-PPT > Format chart > Format Data Point > Series Options >  "Gap Width"
+	 * - width (percent)
+	 * - range: `0`-`500`
+	 * @default 150
+	 */
 	barGapWidthPct?: number
 	barGrouping?: string
+	/**
+	 * MS-PPT > Format chart > Format Data Point > Series Options >  "Series Overlap"
+	 * - overlap (percent)
+	 * - range: `-100`-`100`
+	 * @since v3.9.0
+	 * @default 0
+	 */
+	barOverlapPct?: number
 }
 export interface IChartPropsChartDoughnut {
 	dataNoEffects?: boolean
@@ -1240,11 +1294,43 @@ export interface IChartPropsChartDoughnut {
 }
 export interface IChartPropsChartLine {
 	lineDash?: 'dash' | 'dashDot' | 'lgDash' | 'lgDashDot' | 'lgDashDotDot' | 'solid' | 'sysDash' | 'sysDot'
+	/**
+	 * MS-PPT > Chart format > Format Data Series > Marker Options > Built-in > Type
+	 * - marker type
+	 * @default circle
+	 */
 	lineDataSymbol?: 'circle' | 'dash' | 'diamond' | 'dot' | 'none' | 'square' | 'triangle'
+	/**
+	 * MS-PPT > Chart format > Format Data Series > [Marker Options] > Border > Color
+	 * - border color
+	 * @default circle
+	 */
 	lineDataSymbolLineColor?: string
+	/**
+	 * MS-PPT > Chart format > Format Data Series > [Marker Options] > Border > Width
+	 * - border width (points)
+	 * @default 0.75
+	 */
 	lineDataSymbolLineSize?: number
+	/**
+	 * MS-PPT > Chart format > Format Data Series > Marker Options > Built-in > Size
+	 * - marker size
+	 * - range: 2-72
+	 * @default 6
+	 */
 	lineDataSymbolSize?: number
+	/**
+	 * MS-PPT > Chart format > Format Data Series > Line > Width
+	 * - line width (points)
+	 * - range: 0-1584
+	 * @default 2
+	 */
 	lineSize?: number
+	/**
+	 * MS-PPT > Chart format > Format Data Series > Line > Smoothed line
+	 * - "Smoothed line"
+	 * @default false
+	 */
 	lineSmooth?: boolean
 }
 export interface IChartPropsChartPie {
@@ -1259,7 +1345,12 @@ export interface IChartPropsChartPie {
 	firstSliceAng?: number
 }
 export interface IChartPropsChartRadar {
-	radarStyle?: 'standard' | 'marker' | 'filled'
+	/**
+	 * MS-PPT > Chart Type > Waterfall
+	 * - radar chart type
+	 * @default standard
+	 */
+	radarStyle?: 'standard' | 'marker' | 'filled' // TODO: convert to 'radar'|'markers'|'filled' in 4.0 (verbatim with PPT app UI)
 }
 export interface IChartPropsDataLabel {
 	dataLabelBkgrdColors?: boolean
@@ -1367,6 +1458,8 @@ export interface ISlideRelMedia {
 	path?: string
 	extn?: string
 	data?: string | ArrayBuffer
+	/** used to indicate that a media file has already been read/enocded (PERF) */
+	isDuplicate?: boolean
 	isSvgPng?: boolean
 	svgSize?: { w: number; h: number }
 	rId: number
@@ -1445,9 +1538,8 @@ export interface PresLayout {
 export interface SlideNumberProps extends PositionProps, TextBaseProps {
 	/**
 	 * margin (points)
-	 * TODO: convert to inches in 4.0 (valid values are 0-22)
 	 */
-	margin?: Margin
+	margin?: Margin // TODO: convert to inches in 4.0 (valid values are 0-22)
 }
 export interface SlideMasterProps {
 	/**