+++ src/gen-media.ts
@@ -15,9 +15,23 @@ export function encodeSlideMediaRels(layout: PresSlide | SlideLayout): Promise<s
 	const https = typeof require !== 'undefined' && typeof window === 'undefined' ? require('https') : null // NodeJS
 	let imageProms: Promise<string>[] = []
 
-	// A: Read/Encode each audio/image/video thats not already encoded (eg: base64 provided by user)
-	layout._relsMedia
-		.filter(rel => rel.type !== 'online' && !rel.data && (!rel.path || (rel.path && rel.path.indexOf('preencoded') === -1)))
+	// A: Capture all audio/image/video candidates for encoding (filtering online/pre-encoded)
+	let candidateRels = layout._relsMedia.filter(rel => rel.type !== 'online' && !rel.data && (!rel.path || (rel.path && rel.path.indexOf('preencoded') === -1)))
+
+	// B: PERF: Mark dupes (same `path`) so that we dont load same media over-and-over
+	let unqPaths: string[] = []
+	candidateRels.forEach(rel => {
+		if (unqPaths.indexOf(rel.path) === -1) {
+			rel.isDuplicate = false
+			unqPaths.push(rel.path)
+		} else {
+			rel.isDuplicate = true
+		}
+	})
+
+	// C: Read/Encode each unique audio/image/video path
+	candidateRels
+		.filter(rel => !rel.isDuplicate)
 		.forEach(rel => {
 			imageProms.push(
 				new Promise((resolve, reject) => {
@@ -26,22 +40,26 @@ export function encodeSlideMediaRels(layout: PresSlide | SlideLayout): Promise<s
 						try {
 							let bitmap = fs.readFileSync(rel.path)
 							rel.data = Buffer.from(bitmap).toString('base64')
+							candidateRels.filter(dupe => dupe.isDuplicate && dupe.path === rel.path).forEach(dupe => (dupe.data = rel.data))
 							resolve('done')
 						} catch (ex) {
 							rel.data = IMG_BROKEN
+							candidateRels.filter(dupe => dupe.isDuplicate && dupe.path === rel.path).forEach(dupe => (dupe.data = rel.data))
 							reject('ERROR: Unable to read media: "' + rel.path + '"\n' + ex.toString())
 						}
 					} else if (fs && https && rel.path.indexOf('http') === 0) {
-						https.get(rel.path, res => {
+						https.get(rel.path, (res: any) => {
 							let rawData = ''
 							res.setEncoding('binary') // IMPORTANT: Only binary encoding works
-							res.on('data', chunk => (rawData += chunk))
+							res.on('data', (chunk: string) => (rawData += chunk))
 							res.on('end', () => {
 								rel.data = Buffer.from(rawData, 'binary').toString('base64')
+								candidateRels.filter(dupe => dupe.isDuplicate && dupe.path === rel.path).forEach(dupe => (dupe.data = rel.data))
 								resolve('done')
 							})
-							res.on('error', ex => {
+							res.on('error', (_ex: any) => {
 								rel.data = IMG_BROKEN
+								candidateRels.filter(dupe => dupe.isDuplicate && dupe.path === rel.path).forEach(dupe => (dupe.data = rel.data))
 								reject(`ERROR! Unable to load image (https.get): ${rel.path}`)
 							})
 						})
@@ -53,6 +71,7 @@ export function encodeSlideMediaRels(layout: PresSlide | SlideLayout): Promise<s
 							let reader = new FileReader()
 							reader.onloadend = () => {
 								rel.data = reader.result
+								candidateRels.filter(dupe => dupe.isDuplicate && dupe.path === rel.path).forEach(dupe => (dupe.data = rel.data))
 								if (!rel.isSvgPng) {
 									resolve('done')
 								} else {
@@ -69,6 +88,7 @@ export function encodeSlideMediaRels(layout: PresSlide | SlideLayout): Promise<s
 						}
 						xhr.onerror = ex => {
 							rel.data = IMG_BROKEN
+							candidateRels.filter(dupe => dupe.isDuplicate && dupe.path === rel.path).forEach(dupe => (dupe.data = rel.data))
 							reject(`ERROR! Unable to load image (xhr.onerror): ${rel.path}`)
 						}